Partially implements task #81.
One thing before finishing this task is:
There is still a delay before show the "if load" alert dialog, not sure
what costs this time.
Keep thinking...
----------------------
Another opinion is: we should always try to load automatically, if the
performance is not good, then fix the performance bug. A possible
solution might be using adaptor tech for the gallery. Will investigate
later.
--------------------------
Now, just add a switch flag in AppConfig class.